MAINT: signal: consolidate impl files for spline filtering functions
- merge _splines.py and _bsplines.py
- rename the merged implementation file into _spline_filters.py

None of this is user visible.
