+++ mypy.ini
@@ -342,7 +342,7 @@ ignore_errors = True
 [mypy-scipy.signal.windows._windows]
 ignore_errors = True
 
-[mypy-scipy.signal._bsplines]
+[mypy-scipy.signal._spline_filters]
 ignore_errors = True
 
 [mypy-scipy.sparse.linalg._isolve.tests.test_gcrotmk]
