+++ scipy/signal/meson.build
@@ -81,7 +81,7 @@ py3.extension_module('_spline',
 py3.install_sources([
     '__init__.py',
     '_arraytools.py',
-    '_bsplines.py',
+    '_spline_filters.py',
     '_czt.py',
     '_filter_design.py',
     '_fir_filter_design.py',
@@ -94,7 +94,6 @@ py3.install_sources([
     '_signaltools.py',
     '_spectral_py.py',
     '_spline.pyi',
-    '_splines.py',
     '_upfirdn.py',
     '_waveforms.py',
     '_wavelets.py',
