+++ scipy/signal/tests/meson.build
@@ -4,6 +4,7 @@ py3.install_sources([
     'mpsig.py',
     'test_array_tools.py',
     'test_bsplines.py',
+    'test_splines.py',
     'test_cont2discrete.py',
     'test_czt.py',
     'test_dltisys.py',
@@ -17,7 +18,6 @@ py3.install_sources([
     'test_short_time_fft.py',
     'test_signaltools.py',
     'test_spectral.py',
-    'test_splines.py',
     'test_upfirdn.py',
     'test_waveforms.py',
     'test_wavelets.py',
